move some defaults to initConfig
