+++ packages/core/src/artifacts.ts
@@ -8,7 +8,7 @@ import { printGeneratedTypes } from './lib/schema-type-printer'
 import { ExitError } from './scripts/utils'
 import { initialiseLists } from './lib/core/initialise-lists'
 import { printPrismaSchema } from './lib/core/prisma-schema-printer'
-import { initConfig } from './lib/config'
+import { initConfig } from './system'
 
 export function getFormattedGraphQLSchema (schema: string) {
   return (
@@ -93,11 +93,11 @@ export function getSystemPaths (cwd: string, config: KeystoneConfig) {
     : null
 
   const builtTypesPath = config.types?.path
-    ? path.join(cwd, config.types.path)
+    ? path.join(cwd, config.types.path) // TODO: enforce initConfig before getSystemPaths
     : path.join(cwd, 'node_modules/.keystone/types.ts')
 
   const builtPrismaPath = config.db?.prismaSchemaPath
-    ? path.join(cwd, config.db.prismaSchemaPath)
+    ? path.join(cwd, config.db.prismaSchemaPath) // TODO: enforce initConfig before getSystemPaths
     : path.join(cwd, 'schema.prisma')
 
   const relativePrismaPath = prismaClientPath
@@ -105,7 +105,7 @@ export function getSystemPaths (cwd: string, config: KeystoneConfig) {
     : '@prisma/client'
 
   const builtGraphqlPath = config.graphql?.schemaPath
-    ? path.join(cwd, config.graphql.schemaPath)
+    ? path.join(cwd, config.graphql.schemaPath) // TODO: enforce initConfig before getSystemPaths
     : path.join(cwd, 'schema.graphql')
 
   return {
