+++ packages/core/src/lib/config.ts
@@ -1,7 +1,7 @@
 import type { KeystoneConfig } from '../types'
 import { idFieldType } from './id-field'
 
-function applyIdFieldDefaults (config: KeystoneConfig): KeystoneConfig['lists'] {
+export function applyIdFieldDefaults (config: KeystoneConfig): KeystoneConfig['lists'] {
   // some error checking
   for (const [listKey, list] of Object.entries(config.lists)) {
     if (list.fields.id) {
@@ -52,21 +52,4 @@ function applyIdFieldDefaults (config: KeystoneConfig): KeystoneConfig['lists'] 
   }
 
   return listsWithIds
-}
-
-export function initConfig (config: KeystoneConfig) {
-  if (!['postgresql', 'sqlite', 'mysql'].includes(config.db.provider)) {
-    throw new TypeError(
-      'Invalid db configuration. Please specify db.provider as either "sqlite", "postgresql" or "mysql"'
-    )
-  }
-
-  // WARNING: Typescript should prevent this, but empty string is useful for Prisma errors
-  config.db.url ??= 'postgres://'
-
-  // TODO: use zod or something if want to follow this path
-  return {
-    ...config,
-    lists: applyIdFieldDefaults(config),
-  }
 }
